Rewrite debug core to use UUIDs and be nicer to compiler
It turns out there was no need for a `stack` organization, simply
having one name for each capture-method and result is sufficient.
Moreover, now in stepping we can capture without overwriting the original
method, so the method can be used as a key to retrieve the capturing
function. This should be much friendlier to the compiler.

Unfortunately, for stack traces it still seems necessary (or at least
expedient) to overwrite.