Handle "StashingFailed" exceptions
These occur when point is set at a position not reached during
execution of the code. This should be "innocuous" although the user
should be warned about the problem. This sets the proper warnmsg
and reconstructs the input (preserving the location of point)
so that the user can try again.