Manage exceptions when trying to look up methods
The current strategy of printing a full stacktrace is helpful for
debugging but leads to a feeling that this is unfinished. (It's true,
but not just because of the error.) By catching the error and then
explaining the root cause, it seems more useful to users and a sign
of greater polish in the package.